🚚 Move logic folder into src
